Parse vote messages as byte to avoid overflow errors (votes can never be negative or 0, and currently the maximum possible valid vote number is 20, using an entire int is unnecessary)

+++ RiskOfChaos/EffectHandling/Controllers/ChatVoting/ChaosEffectActivationSignaler_ChatVote.cs
@@ -73,7 +73,7 @@ namespace RiskOfChaos.EffectHandling.Controllers.ChatVoting
         protected void processVoteMessage(string userId, string message)
         {
             if (_effectVoteSelection.IsVoteActive &&
-                int.TryParse(message, out int voteNumber))
+                byte.TryParse(message, out byte voteNumber))
             {
                 int voteOptionIndex = getVoteIndex(voteNumber);
